requirements.txt
aloe==0.1.2
appnope==0.1.0
blessings==1.6
coverage==4.0.3
decorator==4.0.9
Django==1.9.5
django-registration==2.0.4
djangorestframework==3.3.3
factory-boy==2.6.1
fake-factory==0.5.7
future==0.15.2
gherkin-official==4.0.0
nose==1.3.7
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.5.2
repoze.lru==0.6
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
